backend/mysite/requirements.txt
Django==5.0.6
# asgiref==3.3.2
sqlparse==0.4.2
# six==1.15.0
psycopg2>=2.9.1
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.2.2
django-allauth==0.52.0
dj-rest-auth[with_social]==4.0.0
# django-cors-headers==4.1.0
